Move more logic into common package.
