+++ pkg/cmd/init/init.go
@@ -3,6 +3,8 @@ package init
 import (
 	"fmt"
 
+	"github.com/jlucktay/stack/pkg/common"
+
 	"github.com/spf13/cobra"
 )
 
@@ -19,18 +21,9 @@ func NewCommand() *cobra.Command {
 	to quickly create a Cobra application.`,
 		Run: func(cmd *cobra.Command, args []string) {
 			fmt.Println("init called")
+			common.InitStack()
 		},
 	}
-
-	// Here you will define your flags and configuration settings.
-
-	// Cobra supports Persistent Flags which will work for this command
-	// and all subcommands, e.g.:
-	// initCmd.PersistentFlags().String("foo", "", "A help for foo")
-
-	// Cobra supports local flags which will only run when this command
-	// is called directly, e.g.:
-	// initCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
 
 	return c
 }
