+++ pkg/cmd/issue/issue.go
@@ -2,8 +2,11 @@ package issue
 
 import (
 	"fmt"
+	"os"
 
 	"github.com/spf13/cobra"
+
+	"github.com/jlucktay/stack/pkg/common"
 )
 
 // NewCommand returns the issue command.
@@ -19,18 +22,16 @@ func NewCommand() *cobra.Command {
 	to quickly create a Cobra application.`,
 		Run: func(cmd *cobra.Command, args []string) {
 			fmt.Println("issue called")
+
+			if len(os.Args[2:]) == 0 {
+				fmt.Println("No issue text was given!")
+				cmd.UsageString()
+				os.Exit(1)
+			}
+
+			common.CreateIssue(os.Args[2:]...)
 		},
 	}
-
-	// Here you will define your flags and configuration settings.
-
-	// Cobra supports Persistent Flags which will work for this command
-	// and all subcommands, e.g.:
-	// issueCmd.PersistentFlags().String("foo", "", "A help for foo")
-
-	// Cobra supports local flags which will only run when this command
-	// is called directly, e.g.:
-	// issueCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
 
 	return c
 }
