+++ pkg/common/azure.go
@@ -1,4 +1,4 @@
-package main
+package common
 
 import (
 	"bytes"
